Day #36 v5
1) Login success
2) Delete user success
3) User list success

Next Task:
1) Responsible sbadmin2
2) fix route a little bit
3) disable registration

+++ app/Config/Routes.php
@@ -42,7 +42,8 @@ $routes->get('/clipping/index', 'Clipping::index/$1');
 $routes->get('/clipping/form', 'Clipping::form/$1');
 
 $routes->get('clipping/(:segment)', 'Clipping::detail/$1');
-$routes->delete('/clipping/(:num)', 'Clipping::hapus/$1');
+$routes->delete('/clipping/(:num)', 'Clipping::hapus/$1', ['filter' => 'role:superadmin']);
+$routes->delete('/admin/(:num)', 'Admin::hapus/$1', ['filter' => 'role:superadmin']);
